requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-groq
chromadb
pypdf
python-dotenv
sentence-transformers